requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.57.0
django-ckeditor==6.7.0
django-taggit==4.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
Pillow==10.1.0
django-extensions==3.2.3

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0
coverage==7.3.2

# Production dependencies (commented for MVP)
# gunicorn==21.2.0
# psycopg2-binary==2.9.9
# whitenoise==6.6.0
